// src/app/pages/acopiadores/interface/acopiadores.interface.ts
export interface AcopiadoresConApicultor{
    idProveedor: number;
    nombreProovedor: string;
    tipo: string;
    idDatosFiscales: number;
    idDireccion: number;
    idSagarpa: string | null;
    tipoDeMiel: number;
    empresa: number;
    cantidad: string;
    idEstado: number;
    latitud: number | null;
    longitud: number | null;
    deleteProve: number;
    NombreApicultor: string | null;
    NoColmenas: string;
    NoApiarios: string;
    estatus: number;
}